% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biomed_triangle_plot}
\alias{biomed_triangle_plot}
\title{Plots a triangle plot}
\usage{
biomed_triangle_plot(tri_data)
}
\arguments{
\item{tri_data}{A dataset with the values for each of the three categories.}
}
\value{
The triangle plot.
}
\description{
Same as *triangle_plot*, except that it assumes the categories are ACH (Animal, Cellular/Molecular and Human) and automatically adds the translation axis.
}
