% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{codify_triangle_categories}
\alias{codify_triangle_categories}
\title{Generates codes based on MeSH terms}
\usage{
codify_triangle_categories(terms, ref_table)
}
\arguments{
\item{terms}{A character vector with terms (multiple terms in the same element separated by semicolon).}

\item{ref_table}{A reference table with terms and their categories.}
}
\value{
A vector with the categories of the terms in each element.
}
\description{
Compares a vector of terms (multiple terms separated by semicolons) with a reference table and returns the categories to which the elements belong. The reference table must have two columns, one named Term and another named Category. Categories must be represented by a single letter and each term must belong to a single category.
}
\details{
This is intended as a support function for triangle plots, but can be used for any arbitrary classification based on MeSH terms (or other terms).

Example:
ref_table = data.frame(Term = c("Letter A", "Letter B"), Category = c("A", "B"))
codify_triangle_categories(terms = c("Letter B", "Letter A", "Letter A;Letter B"), ref_table)
Returns: "B"      "A"         "AB"
}
