% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cognitiveCareerPlot}
\alias{cognitiveCareerPlot}
\title{Plots the cognitive career of a researcher}
\usage{
cognitiveCareerPlot(
  M,
  base.size = 10,
  n = 20,
  periods = NULL,
  period.names = NULL
)
}
\arguments{
\item{M}{A bibliometrix dataset}

\item{base.size}{A base size for the nodes in the plot}

\item{n}{The maximum number of papers to include}

\item{periods}{A numeric vector of years to add vertical lines and divide periods}

\item{period.names}{A vector of names for the periods}
}
\value{
A plot of the cogntive career.
}
\description{
The cognitive career plot (REF) is a bibliographic coupling graph of the papers of a single author over the years. This function receives a bibliometrix dataset with the papers of single author and makes such plot.
}
\details{
Periods can be identified (e.g. "before post-doc", "after tenure"). Years which divide periods can be given as a vector. Period names is another vector, with length equal that of the divisions plus one (i.e. if there's one division - length(periods) == 1 - then there's two periods - length(period.names) must be == 2).
}
