% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{collapse_author_names}
\alias{collapse_author_names}
\title{Collapses multiple author names into a single one}
\usage{
collapse_author_names(M, alternatives, unique_name)
}
\arguments{
\item{M}{A bibliometrix dataset.}

\item{alternatives}{A character vector with the multiple names.}

\item{unique_name}{The preferred name into which all other will be turned.}
}
\value{
A character vector, same as AU, but with the names replaced.
}
\description{
When authors have multiple names, you can use this function to search and
replace through the AU field and return a version that collapses all occurrences
into a single one.
}
