% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{count_data_for_triangle}
\alias{count_data_for_triangle}
\title{Counts data for ternary categories}
\usage{
count_data_for_triangle(mesh_categories, ref_categories)
}
\arguments{
\item{mesh_categories}{A vector with categories, as letters (see *codify_triangle_categories*).}

\item{ref_categories}{A character vector with three elements, the three categories.}
}
\value{
A data frame that can be used for triangle plots.
}
\description{
Counts the ocurrences of each category (represented by single letters) in a vector of categories (as returned by *codify_triangle_categories*).
}
\details{
This is intended as a support function for triangle plots. To plot multiple points on the same triangle plot, run this function on grouped/split datasets
}
