% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{extract_countries}
\alias{extract_countries}
\title{Extracts author countries for each paper}
\usage{
extract_countries(M, count_unique = F)
}
\arguments{
\item{M}{A bibliometrix dataset.}

\item{count_unique}{Flag to indicate whether to return an unique list or to include repeated.}
}
\value{
A vector with the countries.
}
\description{
From a M bibliometrix data frame, extracts the author's affiliation countries
from the C1 field.
}
