% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosviewer.R
\name{filter_wos_records}
\alias{filter_wos_records}
\title{Writes a Web of Science record file after excluding DOIs}
\usage{
filter_wos_records(folder, dois_to_exclude)
}
\arguments{
\item{folder}{The folder with the WoS records, in plain text format. Every .txt file will be read.}

\item{dois_to_exclude}{The DOIs to be filtered out.}
}
\value{
Saves the filtered files in a subfolder as a side effect. Returns 0.
}
\description{
This function filters the plain text files with Web of Science records by excluding a designated list of DOIs. It saves the filtered records in the same folder as the original ones, under a subfolder /filtered.
}
\details{
This is helpful for when you want to load the records into R, filter, then have them read by another software (like VOSViewer) that can read the WoS format.

It will read all .txt files in the folder, remove the records associated with the designated DOIs, as well as remove any records without a DOI field (DI).
}
