% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{find_lattes_ID}
\alias{find_lattes_ID}
\title{Finds the matching Lattes ID for a given name}
\usage{
find_lattes_ID(query_name, D)
}
\arguments{
\item{query_name}{The name to be searched.}

\item{D}{The snapshot data table, loaded with read_lattes_snapshot().}
}
\value{
A data table with the matching ID.
}
\description{
Searches for a given name on the loaded Lattes CSV snapshot. To use this function, you must first call read_lattes_snapshot.
Matches are found by calculating Levenshtein distances against the names in the snapshot.
}
