% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{get.biblio.data}
\alias{get.biblio.data}
\title{Obtains the "M" data frame for a generation}
\usage{
get.biblio.data(generations, data_folder)
}
\arguments{
\item{generations}{Integer vector indicating the generation(s) to be loaded.}

\item{data_folder}{Full path to the folder where the data is stored.}
}
\value{
An "M" data frame, as loaded and converted through bibliometrix.
}
\description{
Obtains the "M" data frame using bibliometrix to read all
text files in the folder for the generations specified.
Assumes generations are structured in subfolders named G0, G1, G2 and so on.
}
