% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.R
\name{get.bibliometrix.M}
\alias{get.bibliometrix.M}
\title{Load a M data frame}
\usage{
get.bibliometrix.M(folder, dbsource = "isi", format = "plaintext")
}
\arguments{
\item{folder}{Full path to the folder containing the bibliometric data.}

\item{dbsource}{Passed on to convert2df. Defaults to "isi" (Web of Science).}

\item{format}{Passed on to convert2df. Defaults to "plaintext" (Web of Science).}
}
\value{
An "M" data frame, as loaded and converted through bibliometrix.
}
\description{
Obtains the "M" data frame using bibliometrix to read all
text files in the given folder.
}
