% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{get.full.mesh}
\alias{get.full.mesh}
\title{Queries PubMed to obtain MeSH codes}
\usage{
get.full.mesh(pmids)
}
\arguments{
\item{pmids}{A numeric vector with PubMed IDs.}
}
\value{
A datta frame with the MeSH headings and terms.
}
\description{
Given a list of PubMed IDs, this function queries the PubMed API
(in blocks of 100 PMIDs, with 15 second intervals) for their MeSH
codes.
}
\details{
Codes are returned in a two-column data frame, with the MeSH
headings and complete MeSH terms, separated by semicolons.
}
