% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_path.R
\name{load_citation_net}
\alias{load_citation_net}
\title{Build, prepare and save citation network}
\usage{
load_citation_net(M)
}
\arguments{
\item{M}{A bibliometrix M data frame.}
}
\value{
An igraph citation network.
}
\description{
Helper function: it will call other functions to (1) build the citation net,
(2) simplify it into an acyclic graph and (3) save it in Pajek format (side effect).
}
