% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_path.R
\name{load_main_path}
\alias{load_main_path}
\title{Read, convert and save main path}
\usage{
load_main_path(pajekfile, M, NET, output_path = ".")
}
\arguments{
\item{pajekfile}{The path to the Pajek exported main path file (.net).}

\item{M}{A bibliometrix M data frame.}

\item{NET}{The full igraph network.}

\item{output_path}{Path to a directory to save the main path files.}
}
\value{
An bibliometrix data frame with only the main path nodes.
}
\description{
Helper function: it will call other functions to (1) read the
Pajek exported main path network file (2) save it as a GML
network file and (3) save the list of papers in the main path,
as well as the full Web of Science record for the papers in
the main path.
}
