% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_biblio_network}
\alias{plot_biblio_network}
\title{Plots a bibliographic network}
\usage{
plot_biblio_network(
  M,
  edge_type,
  node_type,
  n = 10,
  title = "",
  layout = "auto",
  cluster_method = "none",
  label_size = 4,
  edge_size = c(0.2, 1),
  node_size = c(2, 10),
  aes_color = NULL,
  min_degree = 1,
  get_network = FALSE
)
}
\arguments{
\item{M}{A bibliometrix M dataframe.}

\item{edge_type}{The type of entity that edges represent (passed to biblioNetwork). Can be "collaboration", "coupling", "co-occurrences" or "co-citation".}

\item{node_type}{The type of entity that nodes represent (passed to biblioNetwork). Can be "authors", "references", "sources", "countries","keywords", "author_keywords", "titles", or "abstracts".}

\item{n}{Number of nodes to be included in the plot (the ones with the highest degrees are selected).}

\item{title}{The title for the plot.}

\item{layout}{The layout type for the network (passed to ggraph::create_layout). Can be "auto" (default), "dendrogram", "manual", "linear", "matrix", "treemap", "circlepack", "partition", "hive".}

\item{cluster_method}{The method for clustering (passed to bibliometrix::networkPlot). Can be "none" (default), optimal", "louvain","infomap","edge_betweenness","walktrap", "spinglass", "leading_eigen" or "fast_greedy".}

\item{label_size}{A numeric vector with the range (min, max) of size for labels, passed to the ggplot2::scale_* function.}

\item{edge_size}{A numeric vector with the range (min, max) of size for edges, passed to the ggplot2::scale_* function.}

\item{node_size}{A numeric vector with the range (min, max) of size for nodes, passed to the ggplot2::scale_* function.}

\item{min_degree}{Minimum number of edges a node must have to be included.}

\item{get_network}{Whether to return the network as well as the plot. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Wrapper around bibliometrix::biblioNetwork and bibliometrix::networkPlot, to be able to change the appearance of the network.
}
