% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts}
\alias{plot_counts}
\title{Plots counts as a barplot}
\usage{
plot_counts(df, bar_color, ylab, title)
}
\arguments{
\item{df}{A data frame with two columns.}

\item{bar_color}{The color of the bars.}

\item{ylab}{The label for the count axis.}

\item{title}{The title for the plot.}
}
\value{
A ggplot object.
}
\description{
Plots a table as a bar plot. The first column contains the names, the second contains the counts.
}
