% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_keywordcloud}
\alias{plot_keywordcloud}
\title{Plots a wordcloud of keywords}
\usage{
plot_keywordcloud(
  M,
  title,
  n_words = 70,
  type = "keywords",
  excluded_terms = c()
)
}
\arguments{
\item{M}{A bibliometrix M data frame.}

\item{title}{The plot title.}

\item{n_words}{Number of words to include (most frequent ones).}

\item{type}{Which keywords to use. Can be "keywords" to use the ID field or "mesh" to use the MeshHeadings field (obtained from PubMed beforehand).}

\item{excluded_terms}{Terms to exclude from the word cloud. Will be converted to uppercase. Default is empty. If you pass "mesh*", it will exclude c("Humans","Male","Female","Adult").}
}
\value{
A ggplot object.
}
\description{
Uses ggwordcloud to plot a word cloud from the ID field from Web of Science (Keywords Plus).
}
