% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rpys}
\alias{plot_rpys}
\title{Plots RPYS}
\usage{
plot_rpys(df, bar_color, title)
}
\arguments{
\item{df}{A bibliometrix M dataset.}

\item{bar_color}{The color of the bars.}

\item{title}{The title for the plot.}
}
\value{
A ggplot object.
}
\description{
RPYS is the Referenced Publication Years Spectroscopy (see https://doi.org/10.1002/asi.23089).
It is a histogram of the years of the cited references for a literature corpus.
}
