% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{read_lattes_snapshot}
\alias{read_lattes_snapshot}
\title{Reads the CSV snapshot from Lattes}
\usage{
read_lattes_snapshot(fn)
}
\arguments{
\item{fn}{The path to the CSV snapshot.}
}
\value{
A data table with the IDs and full names.
}
\description{
The resulting data frame is used by other functions (e.g. find_lattes_ID).
It uses data.table::fread to read the large file. The current snapshots can be obtained at http://vision.ime.usp.br/~jmena/base-2020-marco/
}
