% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosviewer.R
\name{read_vos_cluster_gml}
\alias{read_vos_cluster_gml}
\title{Reads a GML file exported from VOSViewer}
\usage{
read_vos_cluster_gml(gml_file)
}
\arguments{
\item{gml_file}{A GML network file, as exported by VOSViewer.}
}
\value{
A data frame with the labels, DOIs and cluster number.
}
\description{
This is a haphazard way of bringing VOSViewer clustering data back into R. This is helpful to make networks in VOSViewer and bring it back into R (I'm pretty sure that's what bibliometrix does, but I want to do it visually).
}
\details{
It assumes the network is bibliographic coupling of papers.
}
