% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{run_scriptLattes}
\alias{run_scriptLattes}
\title{Runs scriptLattes}
\usage{
run_scriptLattes(SL_path, ID_list, save_xml = "~")
}
\arguments{
\item{SL_path}{Path to scriptlattes main python script.}

\item{ID_list}{A character vector containing the Lattes IDs to be retrieved.}

\item{save_xml}{The folder to output the files.}
}
\value{
If successful, 0. Otherwise, 1.
}
\description{
Runs the scriptLattes tool for a given set of IDs. Requires an internet connection.
}
