% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{scrape_formacao_from_lattes}
\alias{scrape_formacao_from_lattes}
\title{Scrapes data from the Formação section of a Lattes CV}
\usage{
scrape_formacao_from_lattes(fn)
}
\arguments{
\item{fn}{The path to the XML CV file.}
}
\value{
A data frame with the information extracted.
#' @importFrom magrittr %>%
}
\description{
Reads the XML from a Lattes CV (as downloaded by ScriptLattes) and extracts the data in the Formacao section - advisor, institution, year and so on.
}
