% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{str_clean_from_titles}
\alias{str_clean_from_titles}
\title{Cleans the advisor field from Lattes}
\usage{
str_clean_from_titles(s)
}
\arguments{
\item{s}{The string extracted from the advisor field.}
}
\value{
The string with the titles/words removed.
}
\description{
Removes common words added to the advisor field on CV Lattes (e.g. "Dr" or "Professor"). Meant for internal use.
}
