% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{str_flatten_name}
\alias{str_flatten_name}
\title{Flattens the name, removing diacritics}
\usage{
str_flatten_name(s)
}
\arguments{
\item{s}{The string to be flattened.}
}
\value{
The string with the non-marked versions of the vowels.
}
\description{
Flattens the name, removing diacritics
}
