% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattes.R
\name{str_namify}
\alias{str_namify}
\title{Trims and makes Title Case}
\usage{
str_namify(s)
}
\arguments{
\item{s}{The character string to make look like a name.}
}
\value{
The string, in Title Case and trimmed.
}
\description{
Trims and makes Title Case
}
