% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{triangle.plot}
\alias{triangle.plot}
\title{Plots a triangle plot}
\usage{
triangle.plot(tri_data)
}
\arguments{
\item{tri_data}{A dataset with the ternary data percentages.}
}
\value{
The triangle plot.
}
\description{
The biomedical triangle plot (see "Identifying translational science within the triangle of biomedicine") is a ternary plot where each corner represents Animal research, Cellular/Molecular research and Human research. This is a way of visualizing translation.
}
\details{
TODO: this function needs to be rewritten to receive the M dataset - with ACH columns - and
itself should prepare the ternary data, you should just indicate the ACH column. This requires
another function to be called beforehand, to convert MeSH into ACH.
}
