% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{triangle_plot}
\alias{triangle_plot}
\title{Plots a triangle plot}
\usage{
triangle_plot(
  tri_data,
  tri_cols,
  tri_labs = NULL,
  plot_translation_axis = F,
  add_density = T
)
}
\arguments{
\item{tri_data}{A dataset with the values for each of the three categories.}

\item{tri_cols}{The name of the columns in the dataset which represent the categories.}

\item{tri_labs}{The name of the axes (if NULL, defaults to the same name as the columns).}

\item{plot_translation_axis}{Whether to plot the line of translation, from AC to H.}
}
\value{
The triangle plot.
}
\description{
The biomedical triangle plot (see "Identifying translational science within the triangle of biomedicine") is a ternary plot where each corner represents Animal research, Cellular/Molecular research and Human research. This is a way of visualizing translation.
}
\details{
This function is generic. The *biomed_triangle_plot* is closer to the plot in the original paper.

Before calling this function, you can generate the appropriately formatted data with *codify_triangle_categories* and *count_data_for_triangle*.
}
